requirements.txt
alpaca-py==0.19.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
statsmodels==0.14.2
scikit-learn==1.5.1
matplotlib==3.9.0
plotly==5.22.0
jupyter==1.0.0
python-dotenv==1.0.1
requests==2.32.3
pre-commit==3.7.1
ruff==0.5.5
black==24.8.0
isort==5.13.2
pytest==8.3.3
